# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://gabomapay.com/api"
        }
    ],
    "paths": {
...
        "/admin/pays/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__PAYS"
                ],
                "summary": "Ajouter un pays",
                "operationId": "dc0b832cf98916bffcadeffc97e9690a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Gabon"
                                    },
                                    "devise": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/pays/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__PAYS"
                ],
                "summary": "Renvoie les informations d'un pays par son ID",
                "operationId": "12d3ee6690296bb2821ed1711bd1be58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/pays/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__PAYS"
                ],
                "summary": "Editer les informations d'un pays",
                "operationId": "befcfcf155f54d46da7b6a2c621dba61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Gabon"
                                    },
                                    "devise": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/pays/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__PAYS"
                ],
                "summary": "Active ou désactive l'état d'un pays par son ID",
                "operationId": "d046673e18fd7bfbbc23b7516b7efde2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/devises/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__DEVISE"
                ],
                "summary": "Ajouter une devise",
                "operationId": "10ac466f39d6aabee29c1aa3ee2dea5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "dollars"
                                    },
                                    "symbole": {
                                        "type": "string",
                                        "example": "$"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/devises/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__DEVISE"
                ],
                "summary": "Renvoie les informations d'une devise par son ID",
                "operationId": "a15a6c00321f1d7a3fc77e04ea780d5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/devises/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__DEVISE"
                ],
                "summary": "Editer les informations d'une devise",
                "operationId": "1e36bb18f757d010a7cf028ab07e0201",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "dollars"
                                    },
                                    "symbole": {
                                        "type": "string",
                                        "example": "$"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/devises/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__DEVISE"
                ],
                "summary": "Active ou désactive l'état d'une devise par son ID",
                "operationId": "7815b07620c572796bd41bbb8a54e62c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}